refactor(security): per-language route-extractor modules behind RouteExtractor (#69)
Modularizes
  route extraction out of security-consistency.ts:
  - New src/drift/route-extractors/: types.ts (RouteInfo, FileMiddleware, ExtractDeps, RouteExtractor), shared.ts
  (comment markers, isCommentLine, inherited*, findHandlerContent, MUTATION_METHODS), and go/js/python/rust.ts each
  exporting a functional RouteExtractor.
  - security-consistency.ts becomes a SupportedLanguage-keyed ROUTE_EXTRACTORS dispatch table; re-exports
  RouteInfo/FileMiddleware so the AST/suppression modules stay untouched.
  - Reuses file.language (removes the langOfFile extension re-derivation).
  - Functional modules per the approved #69 plan (stateless extractors, no classes).

  Behavior-preserving: AST modules, suppression, finding construction, and RouteInfo baseline serialization untouched.
  Full suite outcomes identical before/after.
